moved Puzzle request to be its own component in react
